requirements.txt
numpy==1.18.2
glfw==1.11.0
Pillow==8.0.1
PyOpenGL==3.1.5
